+++ ReleaseNotes.md
@@ -11,6 +11,9 @@
 - cta/CTA#615 - Going to xrdfs xattr API for EOS5 extended attribute tests (EOS >= 5.2.17)
 - cta/CTA#647 - Upgrade eos-5 to eos-5.2.21-1
 
+### Code Quality
+- cta/CTA#575 - Remove rados metrics logging
+
 # v4.10.10.0-1 / v5.10.10.0-1
 
 ### Features